fixedAV model added
this model uses the values fit from the unimodal condition as descriptions of the A and V percepts used in the CI model. Currently the model doesn't really converge, probably because the remaining free parameters (p common and prior sig) don't really have much of an impact on model accuracy. Would probably be improved by adding a bias term that allows for the mean to move. Could also make the mean locations a free parameter and the model would then likely do a good job fitting.