Implemented bias correction
corrected for bias introduced by improperly calibrated eye tracker (linear corrections only). also refactored code somewhat to reduce the number of unnecessary workspace variables created in the master script.

In the future it will probably make sense to nest many of the things that are currently done in the master script into smaller functions, for instance the various model runs.