updating way the saccade pdf is generated
noticed an error where I was making my pdfs in a way that required independent samples from both xa and xv even when the number of causes is one. This is an issue when comparing to data because for these trials I onle have one saccade so sa = sv. Fixed this by keeping the C=1 and C=2 cases separate until after the integration step, and then summing them